Rename _utils.py to _java.py
This class contains Java-related utility functions
and data structures. So let's name it for what it is.
